+++ src/main/java/kitchenpos/application/order/OrderService.java
@@ -1,8 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.application.order;
 
 import com.google.common.collect.Lists;
 import java.util.List;
 import java.util.stream.Collectors;
+import kitchenpos.application.table.OrderTableService;
 import kitchenpos.application.menu.MenuService;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.order.Order;
