+++ src/main/java/kitchenpos/application/product/ProductService.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.application.product;
 
 import java.util.List;
 import kitchenpos.domain.product.Product;
