+++ src/main/java/kitchenpos/ui/TableGroupRestController.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui;
 
-import kitchenpos.application.TableGroupService;
+import kitchenpos.application.tablegroup.TableGroupService;
 import kitchenpos.dto.tablegroup.TableGroupRequest;
 import kitchenpos.dto.tablegroup.TableGroupResponse;
 import org.springframework.http.ResponseEntity;
